GH-4966 Refactor the desktop entry
Changed the desktop file icon lookup, to enable app icon theming.
The Icon line uses the icon name instead of the full path.
For this to work, the icon is moved to a directory specified by the
standard (`/usr/share/icons/hicolor/scalable/apps/`).
As the standard highly suggests, a 48x48 png icon is also added
(rendered from the svg with inkscape).

Added the `StartupWMClass` directive, so that window managers can
associate the created windows with the desktop icon.

Changed the rpm spec file to reflect the aforementioned changes.

Fixes #4966