GH-5166 Adding refresh during Instance Launch
In the login function of LaunchController, added requestRefresh on the current m_accountToUse.
This allows the accountState of accountToUse to be updated before running through what login is meant to depending on the accountState. This is to allow for the cases where a device might be back online after launching the Launcher. Putting this in the tryagain while loop also allows to check if a users account is expired during the login process.